+++ plc4go/protocols/abeth/readwrite/model/CIPEncapsulationConnectionResponse.go
@@ -185,13 +185,13 @@ func (m *_CIPEncapsulationConnectionResponse) GetPlx4xTypeName() string {
 	return "CIPEncapsulationConnectionResponse"
 }
 
-func (m *_CIPEncapsulationConnectionResponse) GetLengthInBits(ctx context.Context) uint16 {
-	lengthInBits := uint16(m.CIPEncapsulationPacketContract.(*_CIPEncapsulationPacket).getLengthInBits(ctx))
+func (m *_CIPEncapsulationConnectionResponse) GetLengthInBits(ctx context.Context) uint64 {
+	lengthInBits := uint64(m.CIPEncapsulationPacketContract.(*_CIPEncapsulationPacket).getLengthInBits(ctx))
 
 	return lengthInBits
 }
 
-func (m *_CIPEncapsulationConnectionResponse) GetLengthInBytes(ctx context.Context) uint16 {
+func (m *_CIPEncapsulationConnectionResponse) GetLengthInBytes(ctx context.Context) uint64 {
 	return m.GetLengthInBits(ctx) / 8
 }
