+++ plc4go/protocols/ads/discovery/readwrite/model/AdsDiscoveryBlockAmsNetId.go
@@ -245,8 +245,8 @@ func (m *_AdsDiscoveryBlockAmsNetId) GetPlx4xTypeName() string {
 	return "AdsDiscoveryBlockAmsNetId"
 }
 
-func (m *_AdsDiscoveryBlockAmsNetId) GetLengthInBits(ctx context.Context) uint16 {
-	lengthInBits := uint16(m.AdsDiscoveryBlockContract.(*_AdsDiscoveryBlock).getLengthInBits(ctx))
+func (m *_AdsDiscoveryBlockAmsNetId) GetLengthInBits(ctx context.Context) uint64 {
+	lengthInBits := uint64(m.AdsDiscoveryBlockContract.(*_AdsDiscoveryBlock).getLengthInBits(ctx))
 
 	// Const Field (amsNetIdLength)
 	lengthInBits += 16
@@ -257,7 +257,7 @@ func (m *_AdsDiscoveryBlockAmsNetId) GetLengthInBits(ctx context.Context) uint16
 	return lengthInBits
 }
 
-func (m *_AdsDiscoveryBlockAmsNetId) GetLengthInBytes(ctx context.Context) uint16 {
+func (m *_AdsDiscoveryBlockAmsNetId) GetLengthInBytes(ctx context.Context) uint64 {
 	return m.GetLengthInBits(ctx) / 8
 }
